+++ rng_tools/src/generators/gen3/wild/mod.rs
@@ -1,5 +1,8 @@
 mod calc_modulo_cycle_count;
 mod cycle_data_calculator;
+mod cycle_range;
+mod game_data;
+mod filter;
 mod generator;
 mod generator_cycle_tests;
 mod generator_tests;
@@ -10,6 +13,9 @@ mod types;
 
 pub use calc_modulo_cycle_count::*;
 pub use cycle_data_calculator::*;
+pub use cycle_range::*;
+pub use filter::*;
+pub use game_data::*;
 pub use generator::*;
 pub use searcher::*;
 pub use types::*;
